+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/init/FirstEntryPoint.java
@@ -67,6 +67,11 @@ public class FirstEntryPoint implements LoggingCustomizer {
   }
 
   @Override
+  public String name() {
+    return "applicationinsights";
+  }
+
+  @Override
   public void init() {
     try {
       if (DEBUG_SIGNED_JAR_ACCESS) {