+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/profiler/config/ProfilerConfiguration.java
@@ -56,8 +56,10 @@ public abstract class ProfilerConfiguration {
   @Nullable
   public abstract String getCollectionPlan();
 
+  @Nullable
   public abstract String getCpuTriggerConfiguration();
 
+  @Nullable
   public abstract String getMemoryTriggerConfiguration();
 
   @Nullable