+++ agent/agent-tooling/src/test/java/com/microsoft/applicationinsights/agent/internal/profiler/config/ProfilerConfigurationTest.java
@@ -13,14 +13,40 @@ public class ProfilerConfigurationTest {
 
   @Test
   void hasBeenConfiguredDetectsDefaultDate() throws JsonProcessingException {
-    ObjectMapper mapper =
-        new ObjectMapper().configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-
     String configStr =
         "{\"id\":\"an-id\",\"lastModified\":\"0001-01-01T00:00:00+00:00\",\"enabledLastModified\":\"0001-01-01T00:00:00+00:00\",\"enabled\":true,\"collectionPlan\":\"\",\"cpuTriggerConfiguration\":\"--cpu-threshold 80 --cpu-trigger-profilingDuration 120 --cpu-trigger-cooldown 14400 --cpu-trigger-enabled true\",\"memoryTriggerConfiguration\":\"--memory-threshold 80 --memory-trigger-profilingDuration 120 --memory-trigger-cooldown 14400 --memory-trigger-enabled true\",\"defaultConfiguration\":null,\"agentConcurrency\":0}";
 
-    ProfilerConfiguration config = mapper.readValue(configStr, ProfilerConfiguration.class);
+    ProfilerConfiguration config = parseConfig(configStr);
 
     Assertions.assertFalse(config.hasBeenConfigured());
+  }
+
+  @Test
+  void parsingCopesWithNullCpuTriggerConfiguration() throws JsonProcessingException {
+    String configStr =
+        "{\"id\":\"an-id\",\"lastModified\":\"0001-01-01T00:00:00+00:00\",\"enabledLastModified\":\"0001-01-01T00:00:00+00:00\",\"enabled\":true,\"collectionPlan\":\"\",\"cpuTriggerConfiguration\":null,\"memoryTriggerConfiguration\":\"--memory-threshold 80 --memory-trigger-profilingDuration 120 --memory-trigger-cooldown 14400 --memory-trigger-enabled true\",\"defaultConfiguration\":null,\"agentConcurrency\":0}";
+
+    ProfilerConfiguration config = parseConfig(configStr);
+
+    Assertions.assertTrue(config.getCpuTriggerConfiguration() == null);
+  }
+
+  @Test
+  void parsingCopesWithNullMemoryTriggerConfiguration() throws JsonProcessingException {
+    String configStr =
+        "{\"id\":\"an-id\",\"lastModified\":\"0001-01-01T00:00:00+00:00\",\"enabledLastModified\":\"0001-01-01T00:00:00+00:00\",\"enabled\":true,\"collectionPlan\":\"\",\"cpuTriggerConfiguration\":\"--cpu-threshold 80 --cpu-trigger-profilingDuration 120 --cpu-trigger-cooldown 14400 --cpu-trigger-enabled true\",\"memoryTriggerConfiguration\":null,\"defaultConfiguration\":null,\"agentConcurrency\":0}";
+
+    ProfilerConfiguration config = parseConfig(configStr);
+
+    Assertions.assertTrue(config.getMemoryTriggerConfiguration() == null);
+  }
+
+  private static ProfilerConfiguration parseConfig(String configStr)
+      throws JsonProcessingException {
+    ObjectMapper mapper =
+        new ObjectMapper().configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+
+    ProfilerConfiguration config = mapper.readValue(configStr, ProfilerConfiguration.class);
+    return config;
   }
 }